Debug: Reset method can only be used by students

+++ core/views.py
@@ -5,14 +5,17 @@ from django.shortcuts import redirect
 @login_required
 def reset(request):
     user = request.user
-    section_items = user.section_items.all()  # for video items
-    quiz_scores = user.quizscore_set.all()
-    quiz_attempts_answers = user.questionattempt_set.all()
+    is_instructor = bool(user.instructor.all())
 
-    user_objects = [section_items, quiz_scores, quiz_attempts_answers]
+    if not is_instructor:
+        section_items = user.section_items.all()  # for video items
+        quiz_scores = user.quizscore_set.all()
+        quiz_attempts_answers = user.questionattempt_set.all()
 
-    for element in user_objects:
-        for item in element:
-            item.delete()
+        user_objects = [section_items, quiz_scores, quiz_attempts_answers]
+
+        for element in user_objects:
+            for item in element:
+                item.delete()
 
     return redirect("home")
